feat : 문제 수정 API 구현
resolves : #38

+++ src/main/java/com/jhr/algoNote/service/ProblemService.java
@@ -10,6 +10,7 @@ import com.jhr.algoNote.domain.tag.Tag;
 import com.jhr.algoNote.dto.CreateProblemRequest;
 import com.jhr.algoNote.dto.ProblemResponse;
 import com.jhr.algoNote.dto.UpdateProblemRequest;
+import com.jhr.algoNote.dto.UpdateProblemResponse;
 import com.jhr.algoNote.repository.ProblemRepository;
 import com.jhr.algoNote.repository.ProblemTagRepository;
 import com.jhr.algoNote.repository.query.ProblemQueryRepository;
@@ -103,37 +104,19 @@ public class ProblemService {
     /**
      * 문제 수정, 수정시 요청자와 문제 작성자가 다르면 예외 발생
      *
-     * @param memberId 수정자 id
+     * @param editorId 수정자 id
      * @param request
      * @return 수정된 문제의 id
      * @throws IllegalArgumentException 작성자가 아닙니다.
      */
     @Transactional
-    public Long edit(@NonNull Long memberId, UpdateProblemRequest request) {
-        //엔티티 조회
-        Member member = memberService.findOne(memberId);
-
-        //태그정보 갱신
-        updateTagList(request.getId(), request.getTagText());
-
-        //문제 수정
-        Problem problem = problemRepository.findById(request.getId());
-        validateWriterAndEditorAreSame(memberId, problem);
-
-        problem.getContent().editText(request.getContentText());
-        //문제 update
-        problem.update(request.getTitle(),
-            request.getSite(),
-            request.getUrl());
-
-        return problem.getId();
+    public Long update(@NonNull Long editorId, @NonNull Long problemId, UpdateProblemRequest request) {
+        validateWriterAndEditorAreSame(editorId, problemRepository.findById(problemId).getMember().getId());
+        update(problemId, request);
+        return problemId;
     }
-
-    /*
-        수정자가 작성자와 동일한지 검증
-     */
-    private void validateWriterAndEditorAreSame(Long memberId, Problem problem) {
-        if (!memberId.equals(problem.getMember().getId())) {
+    private void validateWriterAndEditorAreSame(Long editorId, Long writerId) {
+        if (!editorId.equals(writerId)) {
             throw new IllegalArgumentException("작성자가 아닙니다.");
         }
     }
@@ -143,12 +126,21 @@ public class ProblemService {
     protected boolean updateTagList(Long problemId, String tagText) {
 
         Problem problem = problemRepository.findById(problemId);
-        //기존 태그 삭제
-        problemTagRepository.deleteAllByProblemId(problem);
-        //태그 갱신
-        problem.renewalProblemTag(createProblemTagListWithText(tagText));
 
-        return true;
+        if (!equals(tagText, getTagText(problem.getProblemTags()))) {
+            problemTagRepository.deleteAllByProblemId(problem);
+            problem.renewalProblemTag(createProblemTagListWithText(tagText));
+            return true;
+        }
+        return false;
+    }
+
+    private static boolean equals(String tagText, String originalTegText) {
+        if ((originalTegText.length() == tagText.length()) &&
+            originalTegText.equals(tagText)) {
+            return true;
+        }
+        return false;
     }
 
 
@@ -184,5 +176,14 @@ public class ProblemService {
         return problems.stream().map(p -> new ProblemResponse(p)).collect(toList());
     }
 
+
+    @Transactional
+    public UpdateProblemResponse update(Long id, UpdateProblemRequest request) {
+        updateTagList(id, request.getTagText());
+        Problem problem = problemRepository.findById(id);
+        problem.patch(request.getTitle(), request.getContentText(), request.getSite(),
+            request.getUrl());
+        return UpdateProblemResponse.of(problem,getTagText(problem.getProblemTags()));
+    }
 
 }
